+++ sdk/identity/azure-identity/azure/identity/_credentials/vscode_credential.py
@@ -7,9 +7,10 @@ from typing import TYPE_CHECKING
 from .._exceptions import CredentialUnavailableError
 from .._constants import AZURE_VSCODE_CLIENT_ID
 from .._internal.aad_client import AadClient
-if sys.platform.startswith('win'):
+
+if sys.platform.startswith("win"):
     from .win_vscode_adapter import get_credentials
-elif sys.platform.startswith('darwin'):
+elif sys.platform.startswith("darwin"):
     from .macos_vscode_adapter import get_credentials
 else:
     from .linux_vscode_adapter import get_credentials
@@ -24,6 +25,7 @@ class VSCodeCredential(object):
     """Authenticates by redeeming a refresh token previously saved by VS Code
 
         """
+
     def __init__(self, **kwargs):
         self._client = kwargs.pop("_client", None) or AadClient("organizations", AZURE_VSCODE_CLIENT_ID, **kwargs)
 
@@ -45,9 +47,8 @@ class VSCodeCredential(object):
 
         refresh_token = get_credentials()
         if not refresh_token:
-            raise CredentialUnavailableError(
-                message="No Azure user is logged in to Visual Studio Code."
-            )
-        token = self._client.get_cached_access_token(scopes) \
-                or self._client.obtain_token_by_refresh_token(refresh_token, scopes, **kwargs)
+            raise CredentialUnavailableError(message="No Azure user is logged in to Visual Studio Code.")
+        token = self._client.get_cached_access_token(scopes) or self._client.obtain_token_by_refresh_token(
+            refresh_token, scopes, **kwargs
+        )
         return token