+++ sdk/identity/azure-identity/azure/identity/aio/_credentials/vscode_credential.py
@@ -8,15 +8,18 @@ from .._credentials.base import AsyncCredentialBase
 from ..._constants import AZURE_VSCODE_CLIENT_ID
 from .._internal.aad_client import AadClient
 from ..._credentials.vscode_credential import get_credentials
+
 if TYPE_CHECKING:
     # pylint:disable=unused-import,ungrouped-imports
     from typing import Any, Iterable, Optional
     from azure.core.credentials import AccessToken
 
+
 class VSCodeCredential(AsyncCredentialBase):
     """Authenticates by redeeming a refresh token previously saved by VS Code
 
         """
+
     def __init__(self, **kwargs):
         self._client = kwargs.pop("_client", None) or AadClient("organizations", AZURE_VSCODE_CLIENT_ID, **kwargs)
 
@@ -49,9 +52,7 @@ class VSCodeCredential(AsyncCredentialBase):
 
         refresh_token = get_credentials()
         if not refresh_token:
-            raise CredentialUnavailableError(
-                message="No Azure user is logged in to Visual Studio Code."
-            )
+            raise CredentialUnavailableError(message="No Azure user is logged in to Visual Studio Code.")
         token = self._client.get_cached_access_token(scopes)
         if not token:
             token = await self._client.obtain_token_by_refresh_token(refresh_token, scopes, **kwargs)